requirements.txt
streamlit
openai>=1.0.0
openai-whisper
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
ollama>=0.1.0
python-dotenv>=1.0.0
soundfile>=0.12.1
